# Conflicts:
#	edu.nd.dronology.gstation.python.connector/.classpath
#	edu.nd.dronology.gstation.python.connector/src/edu/nd/dronology/gstation/python/connector/PythonBase2.java
#	edu.nd.dronology.services.core/src/edu/nd/dronology/services/core/util/FileManager.java
#	edu.nd.dronology.services/src/edu/nd/dronology/services/supervisor/WorkspaceInitializer.java

+++ eclipse/edu.nd.dronology.ui.cc.main/src/edu/nd/dronology/ui/cc/main/editor/simulatorscenario/assigndroneviewer/DroneAssignmentViewer.java
@@ -79,8 +79,8 @@ public class DroneAssignmentViewer extends LeftToRightDragAndDropViewer<ISimulat
 		setCordDialog.open();
 
 		try {
-			elem.setStartCoordinate(Long.parseLong(setCordDialog.getLatitude()), Long.parseLong(setCordDialog.getLongitude()),
-					Integer.parseInt(setCordDialog.getAltitude()));
+			elem.setStartCoordinate(Double.parseDouble(setCordDialog.getLatitude()), Double.parseDouble(setCordDialog.getLongitude()),
+					Double.parseDouble(setCordDialog.getAltitude()));
 		} catch (NumberFormatException e) {
 			e.printStackTrace();
 		}
